DTO proveedor,factura y producto

+++ TP1/src/main/java/com/trabajo_practico/gestion_comercial/service/ProveedorService.java
@@ -1,11 +1,14 @@
 package com.trabajo_practico.gestion_comercial.service;
 
+import com.trabajo_practico.gestion_comercial.dto.CreateUpdateProveedorDTO;
+import com.trabajo_practico.gestion_comercial.dto.ProveedorDTO;
 import com.trabajo_practico.gestion_comercial.model.Proveedor;
 import com.trabajo_practico.gestion_comercial.repository.ProveedorRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class ProveedorService {
@@ -16,32 +19,44 @@ public class ProveedorService {
         this.proveedorRepository = proveedorRepository;
     }
 
-    public List<Proveedor> listarProveedores() {
-        return proveedorRepository.findAll();
+    public ProveedorDTO crearProveedor(CreateUpdateProveedorDTO createUpdateProveedorDTO) {
+        var proveedor = generateProveedor(createUpdateProveedorDTO);
+        return new ProveedorDTO(proveedorRepository.save(proveedor));
     }
 
-    public Optional<Proveedor> obtenerProveedorPorId(Long id) {
-        return proveedorRepository.findById(id);
+    public List<ProveedorDTO> obtenerTodos() {
+        return proveedorRepository.findAll().stream().map(ProveedorDTO::new).collect(Collectors.toList());
     }
 
-    public Proveedor crearProveedor(Proveedor proveedor) {
-        return proveedorRepository.save(proveedor);
+    public Optional<ProveedorDTO> obtenerPorId(Long id) {
+        var proveedor = proveedorRepository.findById(id);
+        return proveedor.map(ProveedorDTO::new);
     }
 
-    public Proveedor actualizarProveedor(Long id, Proveedor proveedorDetalles) {
-        return proveedorRepository.findById(id)
-                .map(proveedor -> {
-                    proveedor.setNombre(proveedorDetalles.getNombre());
-                    proveedor.setRuc(proveedorDetalles.getRuc());
-                    proveedor.setTelefono(proveedorDetalles.getTelefono());
-                    proveedor.setEmail(proveedorDetalles.getEmail());
-                    proveedor.setDireccion(proveedorDetalles.getDireccion());
-                    return proveedorRepository.save(proveedor);
-                }).orElseThrow(() -> new RuntimeException("Proveedor no encontrado con id: " + id));
+    public ProveedorDTO actualizarProveedor(Long id, CreateUpdateProveedorDTO updateProveedorDTO) {
+        var proveedor = generateProveedor(updateProveedorDTO);
+        if (proveedorRepository.existsById(id)) {
+            proveedor.setId(id);
+            return new ProveedorDTO(proveedorRepository.save(proveedor));
+        }
+        return null;
     }
 
     public boolean eliminarProveedor(Long id) {
-        proveedorRepository.deleteById(id);
+        if (proveedorRepository.existsById(id)) {
+            proveedorRepository.deleteById(id);
+            return true;
+        }
         return false;
+    }
+
+    private Proveedor generateProveedor(CreateUpdateProveedorDTO dto) {
+        var proveedor = new Proveedor();
+        proveedor.setNombre(dto.getNombre());
+        proveedor.setRuc(dto.getRuc());
+        proveedor.setTelefono(dto.getTelefono());
+        proveedor.setEmail(dto.getEmail());
+        proveedor.setDireccion(dto.getDireccion());
+        return proveedor;
     }
 }
